+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/config/tls/FilePassword.java
@@ -14,34 +14,42 @@ import java.nio.charset.StandardCharsets;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 
+import edu.umd.cs.findbugs.annotations.Nullable;
+
 /**
  * A reference to the file containing a plain text password in UTF-8 encoding.  If the password file
  * contains more than one line, only the characters of the first line are taken to be the password,
  * excluding the line ending.  Subsequent lines are ignored.
  *
- * @param filePath file containing the password,
+ * @param passwordFile file containing the password,
  */
-public record FilePassword(String filePath) implements PasswordProvider {
+public record FilePassword(String passwordFile) implements PasswordProvider {
+
     @JsonCreator
     public FilePassword {}
 
     @Override
     public String getProvidedPassword() {
-        if (filePath == null) {
-            return null;
-        }
-        try (var fr = new BufferedReader(new FileReader(filePath, StandardCharsets.UTF_8))) {
-            return fr.readLine();
-        }
-        catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
+        return readPasswordFile(passwordFile);
     }
 
     @Override
     public String toString() {
         return "FilePassword[" +
-                "filePath=" + filePath + ']';
+                "filePath=" + passwordFile + ']';
+    }
+
+    @Nullable
+    static String readPasswordFile(String passwordFile) {
+        if (passwordFile == null) {
+            return null;
+        }
+        try (var fr = new BufferedReader(new FileReader(passwordFile, StandardCharsets.UTF_8))) {
+            return fr.readLine();
+        }
+        catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
     }
 
 }